pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tutor-stack-content"
version = "0.1.0"
description = "Content service for the Tutor Stack platform"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Ahmed Sarhan", email = "ahmed.sarhan@example.com" }
]
dependencies = [
    "fastapi==0.110.0",
    "uvicorn==0.27.1",
    "pydantic==2.6.1",
    "python-multipart==0.0.9",
    "tutor-stack-auth",
    "google-cloud-storage==2.14.0",
    "google-auth==2.28.0",
    "python-dotenv==1.0.1",
    "google-generativeai==0.7.1",
    "PyMuPDF==1.24.6"
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.0",
    "pytest-asyncio==0.23.5",
    "httpx==0.27.0",
    "pytest-cov==4.1.0",
    "black==24.2.0",
    "isort==5.13.2",
    "flake8==7.0.0",
    "mypy==1.8.0"
]

[tool.hatch.build.targets.wheel]
packages = ["tutor_stack_content"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=tutor_stack_content --cov-report=term-missing"
filterwarnings = [
    "ignore::PendingDeprecationWarning:starlette.*"
]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
exclude = [
    "venv",
    ".venv",
    ".mypy_cache",
    "__pycache__"
] 